Revert original simple example, create new simple example using api
